RPGF-14/Store language code inside of the locbin file (#1)
* Update magic const to be consistent on read and write

* Update writer to include language code

* Update loc bin loader provider to use byte instead of int

* Update LocData.FromBytes to read language code

* Bump package version

+++ Runtime/LocBinLoaderProvider.cs
@@ -5,7 +5,7 @@ namespace RPGFramework.Localisation
 {
     internal static class LocBinLoaderProvider
     {
-        internal static ILocBinLoader GetLocBinLoader(int version)
+        internal static ILocBinLoader GetLocBinLoader(byte version)
         {
             return version switch
                    {
